server/pyproject.toml
[tool.poetry]
name = "vuefinder"
version = "0.1.0"
description = ""
authors = ["kk <kanbang@163.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
werkzeug = "^3.1.2"
fs = "^2.4.16"
pathvalidate = "^3.2.1"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
python-jose = "^3.3.0"
sqlalchemy = "^2.0.36"
python-multipart = "^0.0.17"
bcrypt = "4.0.1"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
